Add jwtsecurity service deployment and filter toggles
The gm.jwtsecurity filter logically relies on the jwtsecurity service.
This commit exposes a top-level field to enable and configure the
service k8s manifests and exposes a listener-level field to
enable the filter on a sidecar proxy. Compared to some other
filters and services, jwtsecurity requires some user input.
The service requires the creation of three secrets.
Some subset of those keys need to exist in the sidecar proxy's
filter configuration object as well.

Finally, note that the filter also requires an egress connection
between the sidecar proxy and the jwtsecurity service. Following
the same pattern as the ext_authz filter, I created a function-like
definition that returns the list of configuration that connect
a proxy to jwtsecurity. It also returns the key of all those
objects so the user can append it to the proxy object
listener and domain keys. This is the best method I could
conceive for minimizing the user's interaction with CUE.