Fixes oidc default mixins, sets default oidc tls option to true
New oidc-authentication (and validation) filter config values
caused a conflict between the backing schema and the default mixin
because the defaults where defined as a defintion which are closed
by default. The solution (and actual correct pattern) is to write
defaults as a struct since structs are open. This allows for us
to omit optional fields.

Also sets the default oidc filter tls options to true. People shouldn't
use oidc over plaintext channels, and also, our filter will not even
allow it in the first place.

Also updates the submodule.