Update defaults
Informed by 300 tenant benchmarks. Note that dashboard and
observables were not loaded, and the datastore sidecar is
anomalous.

Ignore generated manifests
Update mem limits to include 1/2 Gi buffer

---

Update node selector and tolerations

Down but not defeated we found some stability in
using a taint to give control its own node. We have
work to do to identify why we are seeing inconsistencies
without it.

This is a work around to not block updating the
misenformed defaults that existed in 1.8.0. It is
undesirable but until we understand this problem better
it is better we leave it in.

See [sc-35910].

Along with this we will need to instruct users on
how to apply the appropriate taint and label.

kubectl label nodes <node name> greymatter.io/control-node=true
kubectl taint node -l greymatter.io/control-node=true controlensemble=allowed:NoSchedule