Fix double default error in connections.cue caused by internal tls config logic
When internal mtls is enabled, the logic in the domain definition
sets default values for cert paths. In connections.cue, the inbound_socket
definition also sets defaults on the cert paths. CUE cannot determine which
default to choose, and thus crashes with a disjunction error. The solution
was to treat the inbound domain like the normal edge domain and remove it from
the processing logic since its a special case. Note, that this problem cannot occur
within the cluster because its logic checks for an empty instance array and
mesh connections will never use service discovery.